Always report the FTP auto-index results
Whether new episodes are found or not, the FTP auto-index feature
should report its results so that it is obvious to users that the
operation has completed.  This change makes that happen, plus improves
the messaging so that when new episodes are found, their count is
reported.  Finally, this moves feed generation to outside the loop
over new episodes for the sake of performance.