Remove non-admin feed generation
Administrators should have full control of when their RSS feeds are
generated or refreshed, especially as this operation can be somewhat
filesystem-intensive (aggregating potentially thousands of XML
metadata files).  The feed.xml file is already updated after
administrative actions, can be updated via a private cron URL, and can
be updated explicitly and on-demand through the administrative
console.  That is sufficient.  There's simply no reason to allow every
anonymous page view to perform this function.

In other words, PodcastGenerator *should* be able to run in static
mode on a read-only filesystem.