fix(feed_generator.php): correctly insert episode summary and keywords in feed
The equality operations for longdescPG and keywordsPG are turned to non-equality because otherwise
the itunes:summary and itunes:keywords elements are only emitted into the feed when their respective
backing properties are empty. In addition, because of how simplexml_load_file parses the backing
file for an episode, the `<![CDATA[]]>` value is never matched; simplexml_load_file properly parses
CDATA blocks into PHP strings. Therefore the comparisons for longdescPG and keywordsPG are now
against the empty string.

Putting the output of `htmlspecialchars()` inside an XML CDATA block breaks HTML-in-value support in
podcast clients, instead showing the value as text full of HTML tags (as if one did "view source" on
a webpage). As it is, HTML support in clients only seems to exist for description elements (e.g.
`<description>`, `<itunes:summary>`) so it's pointless using CDATA blocks for anything but
shortdescPG and longdescPG.

Closes #364