fix: stop mixing htmlspecialchars() and CDATA
Putting the output of `htmlspecialchars()` inside an XML CDATA block breaks HTML-in-value support in
podcast clients, instead showing the value as text full of HTML tags (as if one did "view source" on
a webpage). As it is, HTML support in clients only seems to exist for description elements (e.g.
`<description>`, `<itunes:summary>`) so it's pointless using CDATA blocks for anything but
shortdescPG and longdescPG.