fix(themes/default): properly determine media type for OG
Forgot to set the value of `$type` needed for the og:audio or og:video
metadata tag; this fixes it by taking the determination code also used
in singleepisode.php. Since we have the MIME media type through this,
let's also set the og:(media-type):type metadata value as well.