Add a French translation
This hasn’t been tested yet in situ, so there may be some noun/verb
issues still, as I haven’t checked the location of every string.

There are a few cases where the original string is sub-optimal, for
instance where it hardcodes a singular or plural instead of using
gettext’s automated plural rule, or when it hardcodes a number, but
those can be fixed later.