fix: support MPEG-4 and Ogg types
Add supported MIME types for MPEG-4 files in plain, AAC, and 3GP formats
and also add MIME for Ogg. These added MIME types are based on the
values in getid3 itself since that is the source of the types used to
check if an uploaded file is valid or not.

Should fix #442