feat: add custom feed tag config file
Since the regular config system can't really handle multi-line
configuration values, we add a new customtags.xml file, like with
buttons and categories, to store custom feed tags. This change provides
functions to get and set the custom tags, and ensures that the XML file
is created on setup or upgrade.