add config setting for enabling custom tags feature
Since this is an advanced feature, add a config setting which defaults
to disabled so that users need to opt in to the functionality. This
commit adds the configuration option, but does not yet gate the display
and update of the fields on the pages which provide the custom tags
functionality.