refactor(episodes_list.php): use getEpisodes()
Switch to getEpisodes() since it will return future episodes for users
who are logged in (who, right now, would only be admins anyway). Because
getEpisodes() returns each episode's data as an associative array
containing an associative array with the actual data, we need to map
from what the function gives us to the actual episode "objects".