feat(feed_generator.php): add custom sorting
Provide the ability to sort by timestamp or by season and episode
number. Because we need to load each episode's XML sidecar to do the
latter, rewrite getEpisodeFiles() to load the sidecars and have
generateRssItem() use that data since we are now throwing it into the
records in the `$files` array.