fix: avoid errors when saving episode categories
Build an array of categories based on post data instead of using
`$_POST` directly to get categories when building the episode XML
sidecar file. This allows us to set category 0 to "uncategorized" when
no categories are selected, and provide empty strings for the others,
avoiding problems with referencing unset values in `$_POST['category']`
and providing those unsets to `htmlspecialchars()`.