refactor(episodes_edit.php): clean up webform
Put field labels in actual <label> tags, and set id attributes on fields
where they are not already present. Make use of CSS to flag required
fields on the form. Reduce section headers to <h4> so that they don't
clash with the page title which is <h3>. And add some helper code to
shorten lines in the HTML as well as deduplicating certain operations.