ci: avoid packing dev dependencies
Run Composer with --no-dev so that we don't include all our development
dependencies in the build artifact. Until we have a need to have dev
dependencies in the build process, the original job for installing deps
is commented out; when both composer jobs are active the second should
simply remove all the dev dependencies, leaving us with a cleaner
artifact.