Changes calls to filemtime to use lstat so symbolic links work
I prefer creating symbolic links in my media folder to use the FTP Auto Indexing feature rather than copying my files directly into the media folder. The only difficulty I've had is that episode's publication date doesn't use the modification time of the symbolic link file itself, but instead uses the modification time of the file that is linked to.

This commit changes every call to filemtime($filename) to lstat($filename)['mtime']. The only difference between the two formulations is that lstat references the symbolic link itself, while `filemtime` follows the link.

The reason I prefer using symbolic links in the media file is that I use the same media for several different PodcastGenerator podcasts, but I need to set the publication dates of a particular media file differently for each podcast I include it in.