docs: add 3.2 release notes
Despite the claim that 3.2 is available, it is not yet at time of this
commit. However, having this written as-is makes things just a little
easier as we get into the release endgame, since I won't need to edit
the release notes document further when we are about to go gold.