fix(feed_generator.php): swap summary tag content
Put the short description in the `<itunes:summary>` tag instead of the
long description, and put the long description, when present, into the
`<description>` tag instead.

This may lead to some confusion for listeners to existing shows which
upgrade from an older version of Podcast Generator, at it will appear
that the short and long descriptions have changed places in their
podcast apps. Blame the lack of any real specification around summaries
and full descriptions/show notes that all apps adhere to.

fixes #522