fix(themes/default): get cover image file from config
The default theme was not updated to support changed filenames for the
show cover art and would continue to use itunes_image.jpg instead of
what the user had most recently set as the cover image. This updates the
theme to make use of the podcast_cover setting instead of hard-coding
the cover image filename.

Fixes #592