fix(setup): load config before including other scripts
Ensure that the global $config is created and populated before
including any other scripts that may rely on it. This prevents a crash
on successful completion of setup.

This is NOT a backport of f0410bdc2f due to the new configuration system
in development for PG 3.3+ but is otherwise identical to that commit.

fixes #665