chore: reconfigure PHP CodeSniffer
Turn off a pair of PSR12 sniffs that complain when
`declare(strict_types=1);` is placed on the same line as the PHP open
tag. Putting the declare statement there looks and feels much better
than it being separated from the open tag. Also, fix the config for the
EOL sniff, as the exclude needs to be wrapped in a rule tag.