refactor(Configuration): improve escape parsing
Use addcslashes/stripcslashes for string escaping on strings which may
have characters we can't normally put in a single-quoted string. We'll
need this to improve serialization of JSON in config.php. The rewrite of
the parsing code also speeds things up by using a single regex where we
previously had two and names capture groups for easier access in code.

Throw in some extra test cases as well.