live_config: Add initial live items admin page
Add a basic admin page that allows us to set the default stream URL and
MIME content type settings. Since we will have additional pages for
managing live items, this page is put into its own menu in the admin nav
bar.

Ensure that the config keys are created when initializing or upgrading a
PG site.