fix: clear anti-CSRF token after check
Avoid crashing admin pages that iterate through form fields to update
site config due to the existence of the `token` key in the POST data.

We shouldn't be blindly accepting the POST data like this, but for now
this at least solves the page crash problem. We can tackle the security
issue with this as improved user management is developed.