feed: add limits on live items in feed
Update feed generation so that not all live items will be emitted in the
RSS feed. All live status items will continue to be put in the feed, but
pending and ended status items can be limited on total number and date
distance.

By default, the two most recent pending items, to a distance of two
weeks ahead, will be shown in the feed. The most recent ended item, up
to one week ago, will also be shown.