fix(liveitems): cast XML tags for ctors
We need to explicitly cast SimpleXMLElement to string in order for the
DateTime constructor to work properly with the tag values. As well, we must
prepend 'urn:uuid:' to the GUID value in order for the Uuid constructor to
work, as it only accepts UUID URNs and curly-wrapped UUID values.