live_edit_form: provide form action URL to view
Instead of adding param after param to the view function, pass a single
"meta" object to it with the non-model data we need to pass to the form.
If we need to provide even more data to the view later on, we don't have
to worry about making the function prototype unwieldy.