episodes_edit: split out model and view
Create a new EpisodeFormModel to hold model code for the episode edit
and upload forms, and episode_edit_form.php for the view code & HTML.
Switch to these for episode_edit.php. We will likely need to make
further changes to support episode_upload.php, so we'll do that in a
separate commit.